requirements.txt
Flask
psycopg2-binary
python-dotenv
SQLAlchemy
Flask-SQLAlchemy
Flask-Migrate
Flask-Login
Flask-WTF
email_validator
gunicorn
pytest
python-dotenv
werkzeug
uuid
Pillow
Flask-Caching
